Rename README_ANALYSIS.MD to README_ANALYSIS.md
##PROJECT OVERVIEW

This analysis was prepared to identify trends and relationships in the data from several crowdfunding campaigns for theatre-related projects and plays. The project outcomes were examined against the campaign launch dates, and individual funding goals. The data is presented visually, with the ultimate goal of drawing actionable insights for future campaigns. 

##ANALYSIS AND CHALLENGES

#* Analysis*

The data used was drawn from a larger dataset of crowdfunding projects. As the focus of this analysis is theatre-related projects and plays, the raw data was organized to include the desired parameters, such as the parent and sub-categories of the campaigns.

First, the 'Theatre' related projects were analyzed based on the campaign launch date. This includes data for plays, as well and musicals and spaces. The